change the collections route to put it in the users route
